replay viewer: add 1/2x playback speed; Space pauses in the league shell (#1)
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that spends one sim tick every other
frame, the chrome grows a 0.5x chip, and the frame JSON's sp field carries
0.5 while it is active (buildStateJson speed is a float now). The league
shell forwards Space down the command channel so play/pause works while the
shell chrome holds focus, matching the board page's existing Space binding.
chrome_common now reads this game's own HNS_WIRE global (the inherited
CTF_WIRE lookup never resolved here), so the engine's speed list drives the
chips; the viewer test pins the patched chrome bytes.

+++ src/hns/broadcast.nim
@@ -341,7 +341,7 @@ proc buildStateJson*(
   sim: SimServer,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  speed: float,
   maxTick: int,
   looping: bool,
   transportEnabled: bool,
